Add files via upload
There has been some major refactoring here. For starters, all functions other than main and start have been placed into the eseentialFunctions.h files. This has also allowed me to add another data point this project can measure, which is altitude!
The altitude function requires a local air pressure value, which has to be sourced from a reliable weather station, either from an airport or a local weather station. As a result, the function can be extremely off at times, so take this value with a grain of salt. To remedy this, I have added a POST function, which allows you to set this value on the fly.